+++ backend/mora/service/orgunit.py
@@ -176,7 +176,6 @@ class OrgUnit(common.AbstractRelationDetail):
 
 RELATION_TYPES = {
     'address': address.Addresses,
-    'it': itsystem.ITSystems,
     'org_unit': OrgUnit,
 }
 
@@ -362,7 +361,7 @@ def get_orgunit(unitid):
     .. sourcecode:: json
 
       {
-        "location:'Overordnet Enhed/Humanistisk fakultet/Historisk Institut'
+        "location": "Overordnet Enhed/Humanistisk fakultet/Historisk Institut",
         "name": "Afdeling for Fortidshistorik",
         "user_key": "frem",
         "uuid": "04c78fc2-72d2-4d02-b55f-807af19eac48",