+++ frontend/src/components/MoEntry/MoItSystemEntry.vue
@@ -4,12 +4,19 @@
       <mo-it-system-picker 
         class="select-itSystem" 
         v-model="entry.itsystem" 
-        :preselected="entry.uuid"
+        :preselected="entry.itsystem && entry.itsystem.uuid"
       />
+
+      <mo-input
+        class="input-itSystem" 
+        v-model="entry.user_key" 
+        :label="$t('input_fields.account_name')"
+        required
+        />
     </div>
 
     <mo-date-picker-range 
-      v-model="entry.validity" 
+      v-model="entry.validity"
       :initially-hidden="validityHidden"
     />
   </div>
@@ -18,9 +25,11 @@
 <script>
   import MoDatePickerRange from '@/components/MoDatePicker/MoDatePickerRange'
   import MoItSystemPicker from '@/components/MoPicker/MoItSystemPicker'
+  import MoInput from '@/components/atoms/MoInput'
 
   export default {
     components: {
+      MoInput,
       MoDatePickerRange,
       MoItSystemPicker
     },
@@ -42,10 +51,15 @@
       entry: {
         handler (newVal) {
           newVal.type = 'it'
-          if (newVal.itsystem !== undefined) newVal.uuid = newVal.itsystem.uuid
           this.$emit('input', newVal)
         },
         deep: true
+      }
+    },
+
+    computed: {
+      nameId () {
+        return 'mo-itsystem-' + this._uid
       }
     },